+++ src/libs/ice/ICEAddress.h
@@ -35,7 +35,7 @@ namespace ice
 
     NetworkAddress(const in6_addr& ip, unsigned short port);
     NetworkAddress(const in_addr& ip, unsigned short port);
-    NetworkAddress(const sockaddr& addr, unsigned addrLen);
+    NetworkAddress(const sockaddr& addr, size_t addrLen);
     NetworkAddress(const NetworkAddress& src);
     ~NetworkAddress();
